Remove UB from dtoa g_strtod
Summary:
There is undefined behavior in `dtoa`'s `g_strtod` function.
Specifically it overflowed a signed integer.
In this one specific loop, write to an unsigned integer instead to avoid the UB.

The previous suppression did not work in fbcode, as there were issues reading from
the blacklist file. With this fix, the blacklist is unnecessary and was removed.

Reviewed By: tmikov

Differential Revision: D15192098

fbshipit-source-id: acdc332dff18ab9feff17afd3d4264c925e89e66