option -ferror-limit to restrict number of emitted errors
Summary:
Similar to other compilers, introduce a command line option to restrict
the maximum number of compilation error. The name of the option was
copied from Clang: `-ferror-limit` (in GCC the same option is called
`-fmax-errors`). It is initialized to 20 by default, and the value 0 is
reserved to mean "unlimited".

This was prompted by an actual test case - a large file containint
garbage, causing the lexer to emit thousands of errors. Printing these
errors was vewry slow, because we do a non-trivial amount of work to
calculate the coordinates of each error and print the offending line.

Implementing this was actually not entirely trivial, since we don't
have a way to terminate the compilation process abruptly. So we used a
multi-pronged approach.

First, we introduce a new parameter for SourceErrorManager - errorLimit.
It simply stops emitting errors after that number is reached. The rest
of the compiler continues to emit them, but they are ignored. We do
however expose an interface, so the compiler can find out this happened,
if it cares.

This alone resulted in a tremendous improvement of the test case, since
we eliminated the most expensive part.

But we did not stop there. It is pointless to continue parsing when we
have reached the maximum error number. We can't abruptly terminate the
compiler, but we fake it by forcing the lexer pointer to EOF. So, as soon
the maximum error count is reached, the next scanned token will be EOF,
which will quickly terminate the parsing process. That may generate a
few extra syntax errors along the way, but we don't care since they will
be ignored.

Lastly, enhance `SimpleDiagHandlerRAII` to set the error limit to one,
thus terminating parsing faster in all cases where we don't need many
errors.

Add a new test.

Reviewed By: willholen

Differential Revision: D15190000

fbshipit-source-id: a6222c35a9b082b149dd7742b24791751edfb201