add MutableHandle::vmcast()/aliasForOutput()
Summary:
We need the ability to construct MutableHandle from a PinnedHermesValue
similarly to how we do it with Handle, particularly in the interpreter
loop. This is what the new vmcast() static method does.

The difference from Handle is, of course, that a MutableHandle doesn't
need to contain a valid value when it is created - that's why it is
mutable and can be assigned to. So, we also need an ability to create
one from a potentially uninitialized PinnedHermesValue. This is what the
aliasForOutput() method does.

In HERMES_SLOW_DEBUG mode, aliasForOutput() initializes the handle to
Invalid to make sure that we never consume the uninitialized value.

Reviewed By: avp

Differential Revision: D13954828

fbshipit-source-id: d16d66ea80aad232d7ea9f2fc6dddb792786aedd