If the initial memory alloc fails, try with a smaller size
Summary:
If the original max heap is too big to be allocated up-front, try to allocate a smaller amount
of memory.
This is better than an alternative scenario, where a low-memory system might fail to allocate at
all and immediately fatal. With this change, they have a chance of getting a smaller heap and
executing some JS.

Note that this diff does *not* change the max heap size that is passed to the GC. This was done
for two reasons:
* It would require re-building and changing the `GCConfig`, which requires changing a couple function
signatures
* An OOM from a reduced heap would say that the JS heap was exhausted, but without changing
the max heap size they'll instead see that the underlying `PreAllocatedStorageProvider` was exhausted

The code for the shrinking was stolen from GenGC.cpp and adapted. They could share the implementation,
but GenGC is mostly dead so I didn't think it was worth trying to work on it.

Reviewed By: kodafb

Differential Revision: D15185761

fbshipit-source-id: f41778ae2415e3aa8723429f4b5c0f8714477bd7