Add JSGeneratorFunction.
Summary:
Functions which are created with `function*` must be `GeneratorFunction`s,
and require their own `CellKind`s in order to have the correct
properties set on them in the future.

Add a `JSGeneratorFunction` class to account for this.

This requires allowing `JSFunction` to be extended, to allow other
functions to have `CodeBlock`s associated with them.

Reviewed By: tmikov

Differential Revision: D14953193

fbshipit-source-id: 88daa2570bc7bbe22d1905adecd38cbb6f19bc3b