Add GeneratorFunction constructor.
Summary:
The GeneratorFunction constructor, while not exposed directly in the
global object, must still allow dynamic creation of generator functions.

As such, add a `GeneratorFunction` constructor, don't put it in the global
object, and make it call the new `createDynamicFunction` function,
which has been extracted from the `Function` constructor.

Note that actually calling `GeneratorFunction` does not work because
the Hermes parser does not support generators yet.

Reviewed By: dulinriley

Differential Revision: D14976581

fbshipit-source-id: 3aa43061e68d1fcdfbb2b3727db32fe94b69d373