Emit and implement generator instructions.
Summary:
CreateGenerator first creates
a GeneratorInnerFunction and then wraps it in a `JSGenerator`.

SaveGenerator needs to point to its corresponding `ResumeGenerator`.
As such, it has standard and long variants.

StartGenerator checks if we are in SuspendedStart and simply continues,
else it jumps to the saved `nextIP`.

ResumeGenerator is only partially implemented in this commit;
it only handles the `Next` instruction for now, which saves the result
to the output register and moves on.

CompleteGenerator sets the state of the current generator to `Completed`,
allowing the user to use `return` in the generator function.

Reviewed By: tmikov

Differential Revision: D14975221

fbshipit-source-id: 66e6475b2902861f1e1f65063da5f5be23dcd43e