Add generator instructions and functions.
Summary:
Add the necessary IR functionality to support generators.

When generating the IR for a generator function, we make two functions:
an outer GeneratorFunction and an inner function.
The outer function calls CreateGenerator on the inner function,
and returns the resultant generator.

The inner function assumes that it can store and retrieve state from
its own closure. As such, it contains instructions to start, save/yield,
and resume generators.

We do not support `yield*` yet.

CreateGenerator:
An instruction to create a generator given a `Function`.
Used by the GeneratorFunction to create the generator which is returned
upon calling it.

StartGenerator:
Always the first instruction executed when an inner function is called.
Restores values for all local variables in the generator,
and jumps to the resume point of the generator if it's suspended;
if it hasn't been started yet, simply continues execution.

SaveAndYield:
Saves necessary state to the closure and yields execution.
In practice, this will save state and use the return opcode to
allow the caller to get the yield result.

ResumeGenerator:
Placed at the start of the block following the SaveAndYield to which it
corresponds. Resumes execution by loading context from the closure,
and then using state stored in the closure, does one of 3 things:
- Continues execution with the result of the instruction being the
  argument to `.next()`.
- Throws a value immediately
- Returns a value immediately

Note that we don't implement the BCGen instructions for this yet;
those will be implemented in a future diff in the stack.

Reviewed By: tmikov

Differential Revision: D14923050

fbshipit-source-id: 4f702e32f1e4684849b5e8756eeed6827446922d