clean up directive scanning
Summary:
Directives are supposed to be simple after parsing: the
ExpressionStatement AST node has an optional attribute "directive",
which indicates that this node is a valid directive. External parsers
like Flow or Acorn correctly set this.

Our handling however was a complete mess:
- Our parser never set that attribute.
- Instead we relied on flags in StringLiteral. These flags weren't set
  or used consistently.
- In SemanticValidator we looked both for the "directive" attribute and
  for expression statements that looked like directives based on our
  flags.
- AST builder also tried to recreate our flags and the directive
  attribute from external AST (which was completely unnecessary because
  external parsers set it correctly).
- Lastly, we didn't handle the case when a directive was followed by a
  closing brace.

Address all of these shortcomings and clean up and simplify the code.

Add a new test.

Reviewed By: xuyueee

Differential Revision: D15213228

fbshipit-source-id: d06c91def3ca0001da2631cfd08b01b37f411caf