Use cached case conversions when bracket matching
Summary:
When matching a bracket in a case-insensitive regex, for example:

    /[A-Z]/i.exec("d")

we have to "canonicalize" the input character ('d') which is a sort of
funky upper-casing. We have a cache for this case conversion that was
not being used. Switch to the cached form.

We can remove the canonicalize() method as it is now unused.

Reviewed By: tmikov

Differential Revision: D15153813

fbshipit-source-id: 91f7c752e84f1bb18673397a748899d46daf50b6