Stop double-canonicalizing when testing for bracket membership
Summary:
When matching a bracket in a case-insensitive regex, for example

    /[a-z]/i.exec("d")

the ES5.1 spec says to canonicalize (uppercase) the input character (here
'd'), and then test all characters that canonicalize to that result. Prior
to this change, Hermes implemented this as written.

Old behavior:

1. Canonicalize "d" to "D"
2. Test "D" for membership in the range 'a-z'
3. Discover pre-canonicalization of "D" (including "d") and test each for membership.

Implement the obvious optimization where we test for membership before
canonicalizing. That is, new behavior:

1. Test "d" for membership in the range 'a-z'
2. Canonicalize "d" to "D" and test "D"
3. Discover any new pre-canonicalizations of "D" (skipping 'd') and test them.

Reviewed By: tmikov

Differential Revision: D15153815

fbshipit-source-id: 199a5469fe785ce6e007294bc7cd3af16f00cdec