Complete parsing of arrow parameters
Summary:
This completes the parsing of arrow function parameters and brings it up
to par with normal functions. Of course arrow function parameters are
implemented via reparsing, so add reparsing support for patterns and
rest element inside the list.

The only tricky part is that PrimaryExpression and SequenceExpression
now need to parse CoverRestElementNode.

Add and update tests.

Reviewed By: avp

Differential Revision: D15216272

fbshipit-source-id: 514a302397d00abb2c5e031dc97da9a1d35bf082