IRGen for "..." in object destructuring
Summary:
Generate code for the "rest property" in object destructuring. This is
actually an ES2018 feature, so we are catching up with the spec...

There is some disagreement about the name of the ESTree node for it. The
ESTree spec just reuses RestElement, while Flow and others use a node
called RestProperty.

In Hermes I decided to reuse RestElement, to follow the spec and avoid
code duplication. That necessitated an adjustment of the AST produced by
Flow.

Add a test.

Reviewed By: avp

Differential Revision: D15043482

fbshipit-source-id: 3d9c8d640ea507dd96ed9b5ce4b33bbf6fdc9a47