add builtins copyDataProperties() and copyRestArgs()
Summary:
Add two new builtins that will be used by ES6 code generation.

    function copyDataProperties(target, excludedItems)
    function copyRestArgs(fromArg)

copyDataProperties() is used by the rest property. The excluded property
list is passed in the form of a literal object with empty values - this
is the most efficient way to do it for now, although theoretically we
would like to just use some efficient internal representation and
perform no allocations.

Let's examine code generation for this ES6 destructuring:

    var {a, b, ...rest} = obj;

We will produce code conceptually equivalent to:

    var tmp = obj;
    var a = tmp.a;
    var b = tmp.b;
    var rest = HermesInternal.copyDataProperties(tmp, {a:0, b:0})

See how the list of excluded properties is efficiently encoded as keys
in an object literal (which will be serialized in an object buffer and
its hidden class cached upon creation). This also enables us to quickly
check for an excluded property by using the literal's property map.

copyRestArgs() is used when generating code for the "rest" function
parameter. Example:

    function foo(a, b, ...rest) {...}

We will generate:
    function foo(a, b) {
        var rest = HermesInternal.copyRestArgs(2);

Note that this is sub-optimal since what we really want is to only
create the JSArray lazily if really needed, similarly to how we handle
`arguments`. That is tracked in T43408487.

Reviewed By: avp

Differential Revision: D15043486

fbshipit-source-id: aae452e1d4480af9ebfe98e333e9e3abf29368d8