Fix validation of object destructuring declarations
Summary:
SemanticValidator wasn't extracting the list of declared identifiers in
destructuring declarations. In the following example it would
incorrectly consider "b" to be the declared variable, while it is in
fact "c":

    var {b:c} = {}

Add a new test for this and update the other affected test (which goes
to show that tests don't magically test for things the programmer didn't
think of).

While we are at it, moved some of the ES6 tests into the correct
directories.

Reviewed By: avp

Differential Revision: D15104053

fbshipit-source-id: 95f4702e4a95d4f8a4771d7c6df0ecce8c6fa725