Esprima test runner to support '.source.js' tests
Summary:
There is a kind of tests in Esprima whose file name ends with ".source.js".
This kind of test has the format of `var source = ...`, and it is meant to be
`eval`ed and the `source` variable is parsed. So we do the same to run
these tests, allowing us to pass more tests. A few are failing and added
to the blacklist.

Reviewed By: tmikov

Differential Revision: D14737727

fbshipit-source-id: fda801ee8b0dac2962207b647a62a9f33d5a1f05