Add facility for introducing extra strings.
Summary: This change allows us to reserve the IDs following those used by the Predefined Strings for other strings.  Precisely which strings are added is controlled by a preprocessor macro indicating a filename, which has a similar format to `PredefinedStrings.def`, but instead of using `STR` as the mapping macro, it uses `RESERVED`.

Reviewed By: ridiculousfish

Differential Revision: D15051081

fbshipit-source-id: 3c9542775d8e63cd799c8ab2bf6e372f976c7c7d