Test to associate Predefined Strings with Bytecode Version
Summary:
The test commits the current set of Predefined strings in their current order
into a ".lock" file and makes sure the lock file is consistent with how the
Predefined Strings are currently defined.

When a new predefined string is added, this test will fail, and explain that
the bytecode version and the lock file also need to be updated.

It also highlights an issue with the way we currently define predefined
identifiers for communicating with the debugger which needs to be fixed before
landing this test (due to an abundance of caution).  My plan for the fix is to
define the Debugger's symbols dynamically at Runtime.

Reviewed By: ridiculousfish

Differential Revision: D15075616

fbshipit-source-id: 4617667035b27c1a7fcc17abf76b2ea9db1f610b