Remove Debugger-specific Predefined Strings
Summary:
Strings used by the `DebuggerInternal` object were being conditionally added to
`PredefinedStrings.def` based on a macro definition.  Recent changes require
that the compiler and runtime agree on the SymbolIDs for Predefined Strings,
this means that we will run into issues if we build bytecode with the macro
disabled in the compiler, and then run it in a version of the runtime built
with the debugger enabled.

To avoid this pitfall, I'm removing the Debugger-specific PredefinedStrings and
adding them as I create the `DebuggerInternal` object.  This happens after
the predefined strings are initialised, so conditionally adding them will not
disturb the assignment of SymbolIDs which the compiler and Runtime need to
statically agree upon.

Reviewed By: avp

Differential Revision: D15214882

fbshipit-source-id: ec314af6e017fe1e7d647e4347ac1dd77b1edef9