Add HoistStartGenerator pass.
Summary:
Some optimization passes may insert extra instructions at the start
of a `GeneratorInnerFunction`.
For resuming generators to work properly, the first instruction must
be a `StartGeneratorInst` (and this is checked by the verifier).

To remedy this problem, add the `HoistStartGenerator` pass, which
finds the `StartGeneratorInst` in a `GeneratorInnerFunction` and moves
it to the start of the function by force.

Run the pass at the end of the pipeline in every instance.

Reviewed By: tmikov

Differential Revision: D15283421

fbshipit-source-id: b5f12f189de135535e1809ec417472544abb2c66