Avoid using StringTableEntry::isIdentifier
Summary:
The field is technically redundant but is used as a matter of convenience in
some parts of the string handling code in the compiler.  It is also used in
runtime asserts.

This diff gets rid of references to the function when creating
`UniquingStringLiteralAccumulator` and `StringLiteralTable` instances, which
needed to know which entries were identifiers but communicated through
`ConsecutiveStringStorage` instances. i.e. the lifecycle of the strings in a
bytecode module followed this pipeline (steps 1 and 2 are optional, if we are
in delta mode):

  1. (`BaseBCProvider`)
  2. (`ConsecutiveStringStorage`)
  3. `UniquingStringLiteralAccumulator`
  4. `ConsecutiveStringStorage`
  5. `StringLiteralTable`
  6. `ConsecutiveStringStorage`

Now these instances communicate directly with each other, using their
`isIdentifier` vectors to communicate the necessary information.

 1. `BaseBCProvider`
 2. `UniquingStringLiteralAccumulator`
 3. `StringLiteralTable`
 4. `ConsecutiveStringStorage`

Reviewed By: kodafb

Differential Revision: D15333878

fbshipit-source-id: aefdaf37b9954fd2475f10bf0e4fe977d429ac4c