don't assert strictness if errors are present
Summary:
There are two ways to record the strictness - either the parser sets it
in the AST node while parsing, or when we are consuming an external AST,
SemanticValidator scans for "use strict" directives and sets it.
SemanticValidator has an assert() to ensure that the two kinds of
strictness, when both are present, agree with each other.

It turns out that there is a corner case where **invalid input** can
cause the two to disagree (see below). So, the fix is to only assert()
if there are no other errors - we only care about correct strictness if
the source is otherwise error free.

This is the example causing the problem:

    function foo(x = function () {}) {
        "use strict";
    }

What is the strictness of the function in the parameter list? Because it
is not very clear, the spec explicitly prohibits this - functions with
non-simple parameters are not allowed to set strictness. We correctly
report that as an error, but the strictness set by the parser and
detected by SemanticValidate remain different.

Add a test.

Reviewed By: avp

Differential Revision: D15332272

fbshipit-source-id: 3fbc6d88e20e9f47e9fc4d908ff82b5e04ccd840