move CreateArgumentsInst to start of prologue
Summary:
Address a bug where we were trying to access `arguments` in the
initializer expression of a parameter, like this:

    function foo(a = arguments) {}

The problem is that CreateArgumentsInst was created lazily when needed
and insert before "entryTerminator", and "entryTerminator" hadn't been
initialized yet at the point of the prologue where we were initializing
the formal parameters.

To fix it, reverse the logic: always insert CreateArgumentsInst in the
start of the prologue, and later delete it if it has no users. It also
turned out that this was the only reason we kept "entryTerminator", so
it became unnecessary and was removed, for some nice simplification.

Update a couple of affected tests and add a new one specifically for
this.

Reviewed By: avp

Differential Revision: D15292219

fbshipit-source-id: 5b88ad659b252bb1cac518ac13c497143d7525be