Remove busted StringPrimitive::createLongLived length-variant
Summary:
There exists a function StringPrimitive::createLongLived which accepts a
length, expecting the user to fill in the contents later. This function
has never actually worked: it ends up converting the length to a char and
then creating a string of length 1 containing that char.

There is no test coverage and it's obviously not being used. Just remove
it.

Reviewed By: kodafb

Differential Revision: D14662008

fbshipit-source-id: 80465ffec524c926b2f095a42f51eae7d8064477