Tool for analysing bytecode page read traces
Summary:
kodafb introduced instrumentation in D15131881 that traces every page access
and logs how much time it took.

The script introduced in this diff takes the data from that trace and produces
an output where each page offset has been replaced by a bytecode section name.

This script expects the trace to be passed in on STDIN as a TSV table.  I
converted to TSV from the format used to log the trace using this script:

```
#! /usr/bin/env ruby
# io_tsv

puts "Page\tTime (us)"
ARGF.each_line do |l|
  /HermesIO: (.*)$/.match(l) do |m|
    m[1].split(',').each do |p|
      /(\d+) \((\d+)us\)/.match(p) do |m|
        puts "#{m[1]}\t#{m[2]}"
      end
    end
  end
end
```

It also requires access to the bytecode file the trace was reading from, as
well as a version of `hbcdump` that is compatible with the bytecode file.  The
latter was added as a parameter to make it easier to compare traces across
bytecode versions.

Reviewed By: kodafb

Differential Revision: D15182553

fbshipit-source-id: d7b216892cdf1c0d3082cbf2a89b67701d3e39d0