Add uniqued StringPrimitive classes
Summary:
Prior to this fix, DynamicStringPrimitive optionally had a SymbolID
trailing after the character contents. This is incompatible with the GC
and so the SymbolID was not being propertly marked during a collection.

This fix does the following:

1. Introduce a new subclass SymbolStringPrimitive which stores a SymbolID
(as an ordinary instance variable).

2. Equip DynamicStringPrimitive with a template parameter bool Uniqued.
DynamicStringPrimitive inherits from either SymbolStringPrimitive or
StringPrimitive, according to this parameter.

3. Introduce new CellKinds `DynamicUniquedUTF16StringPrimitive` and
`DynamicUniquedASCIIStringPrimitive`.

4. Provide proper GC metadata for both of these such that the SymbolID is
properly marked.

Reviewed By: davedets

Differential Revision: D15389485

fbshipit-source-id: c3df209ee054d96a386eb341fafcdd3345d15766