Fix some step-in behaviors
Summary:
The debugger can't step into a accessor function, or any function that is called
by recursively invoking the interpreter. That is because we are using the single
stepping mechanism for these instructions that could recursively call the interpreter.
The debugger uses the single-step mode to interpret one instruction while we
are still in the debugger. So even we set a breakpoint at the beginning of the
accessor function, it will be ignored.

There are many instructions that could call the interpreter recursively. So we
change the default stepping mechanism to "set a breakpoint at the next instruction
and continue executing until we hit a breakpoint again". We will use single-step
mode for instructions that we know that won't call interpreter recursively, also might
jump to a different location other than the next instruction, since it's not easy to
determine the target without interpreting the instruction. Currently these instructions
are `Throw` and `SwitchImm`.

Reviewed By: avp

Differential Revision: D15290700

fbshipit-source-id: df3678ea7ed73e892b0dd9ff7c35481b4c51f7b5