Check overriding builtin methods with computed property names
Summary:
Previously we throw when there is an attempt to override a static builtin
method, but that only covered the named properties. When it's a computed
property name, like `var isArray = 'isArray'; Array[isArray] = 1;`, it will still throw
with a generic error message about overriding a read-only property. Fix that.

Reviewed By: tmikov

Differential Revision: D15112409

fbshipit-source-id: 4903cfd5339e88d4fc08aa7b54b18c826dd7b821