Add a new directive for turning on static builtin optimization
Summary:
Support a new JS directive `'use static builtin'`. This tells the compiler to perform
static builtin optimization. The directive works similarly as the `'use strict'` directive
in JavaScript. But the difference is that its effect is not limited to the function scope
that it's in. So even when it appears in a function, the static builtin optimization will
be turned on for the whole program.

Also modifies the compiler option for static builtins. The option has three modes:
force on, force off, and auto detect. In auto detect mode, turn on static builtin when
there is the directive; in force on/off, turn it on/off regardless of whether the directive
appears or not.

This makes it easier for individual apps to configure this optimization: just add
the directive to one of the source files instead of adding the compiler flag to the
build pipeline.

This helps us enable static builtins in debug builds.

Reviewed By: tmikov

Differential Revision: D15209191

fbshipit-source-id: f77dbfdb0dd68fef283dd908f2538574a4844b24