Use visitor pattern to centralize bytecode segment ordering
Summary:
Prior to this fix, the bytecode file encoding and decoding was quite
ad-hoc and shared no particular structure. Add a new template function
visitBytecodeSegmentsInOrder() which imposes an order and segment list.
Switch encoding and decoding to this new function. This helps ensure that
encoding and decoding stay in sync, and makes it easier to change the
order or segment list.

Bump the bytecode version since we now align all segments at 4 byte
boundaries (previously this was ad-hoc).

Reviewed By: willholen

Differential Revision: D15155801

fbshipit-source-id: 0e9c3da151ffd2136c634e607ca8ce424d91e4a6