createStringFromUtf8 to check for ASCII
Summary:
A test showed that in practice, over 99.9% of strings passed to
createStringFromUtf8() are in fact in ASCII. These appear to be mainly
property names, so this will hold up across locales.

When creating a UTF8 string in JSI, add a check for ASCII
and potentially save the cost of widening.

Reviewed By: kodafb

Differential Revision: D14662005

fbshipit-source-id: f53ee75fe3f233491060341ad897088dd74bf975