hermes jsi to create strings via createEfficient
Summary:
In some cases, the hermes jsi runtime will create a UTF-16 std::string,
which it then passes to Hermes as an ArrayRef. Hermes will then copy that
string, which seems silly since jsi will just discard the string.

Switch to using the createEfficient API to permit jsi to transfer ownership
directly to Hermes, so Hermes can just adopt the string as an
ExternalStringPrimitive (if it meets the length threshold).

Reviewed By: amnn

Differential Revision: D14662007

fbshipit-source-id: 29cda5e0ec40055af675b352384bf3f09b0e5733