BUG: init capture state before emitParameters()
Summary:
Parameter initializers were emitted too early in the function prologue,
before the "capture state" (the captured values of "this", "new.target"
and "arguments") was initialized. A parameter initializer expression
referring to the captured values would cause a crash. This is the second
similar bug found by fuzzing recently and hopefully the last one.

Add a test.

Reviewed By: amnn

Differential Revision: D15406423

fbshipit-source-id: 441d2928c5214bdd45984a5c53561e3ac763a297