do not export llvm::SourceMgr from SourceErrorManager
Summary:
We want to be able to provide better encapsulation and be able to work
around some problematic aspects of llvm::SourceMgr (for example it
performs a linear search in all buffers to determine where a pointer
belongs, which could be a problem with 20,000 buffers).

So, do not expose it, giving us the freedom to reimplement it, if we
wanted to.

Reviewed By: avp

Differential Revision: D15281449

fbshipit-source-id: 9e1bc08ac70e07e5bae8341cdd8447585502b31a