Implement source coordinate translation for input SourceMap
Summary:
Aakash did the work to merge input source map in SourceMapGenerator. This works fine if `-output-source-map` is used. But it does not work if we want the merged line info to be embedded into HBC bundle.

This diff parses input source map from `-source-map` command line option and does the translation in `SourceErrorManager` using the new `ICoordinateTranslator` interface. This should work for both scenarios(with future diffs).

Reviewed By: tmikov

Differential Revision: D15381318

fbshipit-source-id: 78b8c3889f07bf164f4903a2b3fbbbf48406c171