Support zip base bytecode file
Summary:
In delta mode, we want to be able to take in a base bytecode package (in zip or directory form) that contains multiple segments, and use them to compile some new bytecode that optimizes for delta updates.
This changes assumes that in the zip file, there will be a json file named manifest.json that describes the bytecode files in the zip. We create a base bytecode map from segment id to each deserialized bytecode, and use them to compile new bytecode.

Reviewed By: avp

Differential Revision: D15547865

fbshipit-source-id: f867720f283cf09476a6a8df6163533262692252