Switch tryGetOwnNamedDescriptorFast to return OptValue<bool>
Summary:
Prior to this fix, tryGetOwnNamedDescriptorFast() would return true if a
property was found, false if the property was not found OR if the search
could not be conducted because the property map was missing.

This distinguishes the last two cases, returning None for not-searched and
false for definitely-not-found. The new bool specialization of OptValue
allows for collapsing hasValue and getValue checks into a single comparison,
so this remains equally efficient.

This is required to determine when it is safe to use the prototype cache,
introduced in the next commit. It also makes the best-effort nature of
the function more obvious.

Reviewed By: amnn

Differential Revision: D15235102

fbshipit-source-id: 12afa837520d77f2ee1a354284808bbcb0d886fd