Specialize OptValue for bool
Summary:
Prior to this fix, an OptValue<bool> was implemented as a pair of
booleans. Specialize this with an int, storing -1 to indicate a missing
value. This is important to avoid regressing the GetByID path in the next
commit.

Example improvement:

    bool test(OptValue<bool> v) {
        return v && *v;
    }

is now compiled with a simple >0 check:

    test(OptValue<bool>):
        test    edi, edi
        setg    al
        ret

instead of multiple compares:

    test(OptValue<bool>):
        movzx   eax, di
        cmp     eax, 255
        seta    cl
        test    dil, dil
        setne   al
        and     al, cl
        ret

ARM improves as well, though less dramatically.

Reviewed By: amnn

Differential Revision: D15235103

fbshipit-source-id: e16fa79ec19762c98dc39e141254a826f9bfd775