Allow populating hidden class read caches with prototypes
Summary:
This implements the following hypothesis: if we have a hidden class cache
miss, the value is probably found as an "own property" on the prototype of
the object. It is aimed at code like so:

    f.apply(stuff)

where 'apply' is on the prototype of f.

This change allows the read cache to be populated with some member of the
prototype chain of the object. After testing the cache for the object's hidden
class, we must then search the object's own property storage. If that search
fails, before walking the prototype chain, we check the cache against the
object's immediate prototype. On a cache hit, we can skip the prototype chain
walk.

Running fb4a_marketplace_big_gc_synth:

before:

     1279267 vm - NumGetByIdProto: Number of property 'read by id' in the prototype chain

after:

       29319 vm - NumGetByIdProto: Number of property 'read by id' in the prototype chain
     1249948 vm - NumGetByIdProtoHits: Number of property 'read by id' cache hits for the prototype

so we elide ~97% of prototype chain hits.

Reviewed By: tmikov

Differential Revision: D15195141

fbshipit-source-id: ec57eb8e3ee78c602362d0b49d65871432388320