Fix handle stack limit violation.
Summary:
When we construct a BoundFunction, there's a loop that creates a handle for each argument.  With enough arguments, this can violate the Handle stack limit.  Thus, put a GCScope in the body of the loop, to delete each handle after its use is finished.

Alternatives:
  * We could have a single mutable handle.
  * We could test for the presumably common case in which the number of arguments is reasonable, say <= 8, and have two copies of the loop: one without the GCScope, if the total number of arguments is small, and one with ithe GCScope, if the number is large.
  * Or doubly nest the the loop, in chunks of 16 or 32, with a scope for each outer iteration.

But I assumed that this is not perf-sensitive, so just putting in the GCScope seems like the most straightforward solution.

Reviewed By: tmikov

Differential Revision: D15449007

fbshipit-source-id: 11794aac5f88b010d65a6055226faaac533d1056