Fix bug exposed by a test case from Stephen Fewer.
Summary: Stephen Fewer crafted a program that exposed violation of an implicit invariant we were assuming in the GC's OldGen: that the size() of the generation is >= the "levelOffset()".  See the task for full discussion; the short form is that we need to avoid shrinking the OG size below this limit (since this is the amount of storage actually currently required by the OldGen).

Reviewed By: amnn

Differential Revision: D15449033

fbshipit-source-id: 40778a53a1d9cc31f6003c16e59791e564a7a313