use a callback for finally handler emission
Summary:
When inlining code for "finally" handlers, invoke a callback instead of
compiling a specified AST node. Pass a flag to the handler informing it
whether it is handling "continue" or "break". All of this gives us the
flexibility to do things things which cannot be expressed in JavaScript.
It will be used in the followup commits.

Reviewed By: avp

Differential Revision: D15275509

fbshipit-source-id: 2036ca7efd799558f054cd7c98e0c30ec1cf3b99