Exception handling and iterator closing in for-of
Summary:
Implement the correct spec-compliant behavior in `for-of` when exceptions
occur or the loop terminates prematurely. Basically, when an iterator
is no longer needed and it hasn't "naturally" reached its end by
reporting "done = true", it needs to be closed by calling its optional
method "iterator.return()".

This needs to happen if the for-of loop is terminated prematurely with
an exception or break/return. Additionally, if the "return()" method
throws while we already are handling an exception, the new exception has
to be ignored. Lastly, the "return()" method must always to return an
object and we have to throw TypeError if it didn't.

All of this does make sense, but unfortunately it leads to complications
in the generated code. This commit attempts to save some space by
emitting exception handling code which is impossible to emulate with
JavaScript constructs.

Basically we are creating a synthetic "try/catch" statement covering the
parts of the for-of loop which need to be covered (which is not all of
it). In pseudocode it looks something like this:

    for-of-loop {
        try {
            loop body
        } catch (e) {
            var ret = iterator.return
            if ret !== undefined {
                try {
                    ret();
                catch (ignored) {}
            }
            throw e; // rethrow the exception.
        }
    }

If the inside of the loop executes a break or return, we emit the
following code *outside* of the area covered by the try/catch:

    var ret = iterator.return
    if ret !== undefined {
        var tmp = ret();
        HermesInternal.ensureObject(tmp);
    }

A bunch of tests were added in two versions: one with an explicit
custom iterator and one with a generator.

Reviewed By: avp

Differential Revision: D15275507

fbshipit-source-id: 6bfee779fec6d75aa58d956b58a0498e74b9886e