Rename functions that could recursively call the interpreter
Summary:
There are function calls in the interpreter function that could invoke the interpreter
recursively, like `JSObject::getNamed`. The stack frame created from these kind of
recursive interpreter calls won't save the caller's ip in the `savedIP` register in the
register stack. This causes problem for the debugger when we want to step out of
a function because without caller's IP, we don't know where to step out to.

This diff stack is a fix for the problem. This diff first rename all the functions of this
kind, so when people call these functions in the interpreter, they are reminded of
the nature of these functions. "RJS" means "re-entry into JS" (credit to tmikov).

Reviewed By: tmikov

Differential Revision: D15483208

fbshipit-source-id: 148ff367422dffdc1a8c206e8fd6cd257bfdc84c