Fix source locations for Program and File
Summary:
Previously the end of the source locations for "Program" and "File" nodes
would include the comments at the end, while the beginning of the source
location don't include the comments at the beginning. It's not consistent.
Fix this to always include comments in the "File" node, and always exclude
comments in the "Program" node.

Reviewed By: tmikov

Differential Revision: D15637156

fbshipit-source-id: e5fa87b4e7c0c2b17e9d63b81387470c66dea62b