Back out "[Hermes] Stop using createEfficient in jsi"
Summary:
Put back the call to createEfficient() which creates strings more efficiently.

This had to be backed out because of weak symbol strangeness with u16string.
Now that hermes is statically linked (see parent) it is safe to restore this.

Note despite the middling lab results here, the back-out diff D15525986 was identified
as a significant regression, so I would like to restore this.

Original commit changeset: e03f1eaa77c4

Reviewed By: avp

Differential Revision: D15632772

fbshipit-source-id: f24a38d37c5b26d372b62f756baf636a6285f465