Define allocStack out of line
Summary:
The Runtime::allocStack() function is defined inline. This allows constants
to be propagated into the initValue parameter, which then enables clang to
replace the uninitialized_fill function with a call to memset_pattern_16.

This ends up being a loss on Mac / iOS because our stack frames are small.
memset_pattern_16 is dynamically linked, so every stack allocation
resulted in an indirect call, vectorized loop with scalar prefix, etc.

Define it out of line to defeat this optimization. Also remove the return
value which was unused.

This improves perf for all JS function calls on these platforms. For example
earley-boyer octane benchmark improves by ~10% (score from 3828 to 4194)

Reviewed By: tmikov

Differential Revision: D15679910

fbshipit-source-id: 06b681a645c08ff2e105603f584a3f9905548917