Standardising asserts before calls to madvise
Summary:
kodafb pointed out to me that the linux implementation of `madvise` requires
page-aligned addresses in D15588898 (unlike the BSD implementation of `madvise`
that comes with macOS).

This caused me to realise that we were not being consistent in the way we
tested this pre-condition in OSCompat.  Some places we were asserting it and in
other places we were not.  We also had an assert in `vm_unused` that the length
should also be page aligned but this is not the case.

This diff makes sure that calls to `madvise` in `OSCompatPosix.cpp` assert for
non-page-aligned addresses but not for lengths.

Reviewed By: kodafb

Differential Revision: D15608825

fbshipit-source-id: 25df86fc5b3b2873b28835c627fad432603a68e3